cmd,groot: remove rtree.{Tree,}Scanner
Updates #773.
