+++ groot/rtree/chain_test.go
@@ -6,7 +6,6 @@ package rtree_test
 
 import (
 	"fmt"
-	"io"
 	"reflect"
 	"testing"
 
@@ -352,7 +351,7 @@ func TestChainOf(t *testing.T) {
 	}
 }
 
-func TestChainScanStruct(t *testing.T) {
+func TestChainReaderStruct(t *testing.T) {
 	files := []string{
 		"../testdata/chain.1.root",
 		"../testdata/chain.2.root",
@@ -422,421 +421,26 @@ func TestChainScanStruct(t *testing.T) {
 		return data
 	}
 
-	sc, err := rtree.NewTreeScanner(chain, &Data{})
+	var data Data
+	r, err := rtree.NewReader(chain, rtree.ReadVarsFromStruct(&data))
 	if err != nil {
 		t.Fatal(err)
 	}
-	defer sc.Close()
+	defer r.Close()
 
-	for sc.Next() {
-		var d1 Data
-		err := sc.Scan(&d1)
-		if err != nil {
-			t.Fatal(err)
-		}
-		i := sc.Entry()
-		if !reflect.DeepEqual(d1, want(i)) {
-			t.Fatalf("entry[%d]:\ngot= %#v\nwant=%#v\n", i, d1, want(i))
-		}
-
-		var d2 Data
-		err = sc.Scan(&d2)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if !reflect.DeepEqual(d2, want(i)) {
-			t.Fatalf("entry[%d]:\ngot= %#v\nwant=%#v\n", i, d2, want(i))
+	err = r.Read(func(ctx rtree.RCtx) error {
+		i := ctx.Entry
+		if !reflect.DeepEqual(data, want(i)) {
+			return fmt.Errorf("entry[%d]:\ngot= %#v\nwant=%#v\n", i, data, want(i))
 		}
 		total.got++
-	}
-
-	if err := sc.Err(); err != nil && err != io.EOF {
+		return nil
+	})
+	if err != nil {
 		t.Fatal(err)
 	}
 
 	if total.got != total.want {
-		t.Fatalf("entries scanned differ: got=%d want=%d", total.got, total.want)
-	}
-}
-
-var SumF64 float64 // global variable to prevent unwanted compiler optimization
-
-func BenchmarkChainTreeScannerStruct(b *testing.B) {
-	files := []string{
-		"../testdata/chain.1.root",
-		"../testdata/chain.2.root",
-	}
-
-	trees := make([]rtree.Tree, len(files))
-	for i, fname := range files {
-		f, err := riofs.Open(fname)
-		if err != nil {
-			b.Fatalf("could not open ROOT file %q: %v", fname, err)
-		}
-		defer f.Close()
-
-		obj, err := f.Get("tree")
-		if err != nil {
-			b.Fatal(err)
-		}
-
-		trees[i] = obj.(rtree.Tree)
-	}
-	chain := rtree.Chain(trees...)
-
-	type Data struct {
-		Event struct {
-			Beg       string      `groot:"Beg"`
-			F64       float64     `groot:"F64"`
-			ArrF64    [10]float64 `groot:"ArrayF64"`
-			N         int32       `groot:"N"`
-			SliF64    []float64   `groot:"SliceF64"`
-			StdStr    string      `groot:"StdStr"`
-			StlVecF64 []float64   `groot:"StlVecF64"`
-			StlVecStr []string    `groot:"StlVecStr"`
-			End       string      `groot:"End"`
-		} `groot:"evt"`
-	}
-
-	sc, err := rtree.NewTreeScanner(chain, &Data{})
-	if err != nil {
-		b.Fatal(err)
-	}
-	defer sc.Close()
-
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		err := sc.SeekEntry(0)
-		if err != nil {
-			b.Fatal(err)
-		}
-		for sc.Next() {
-			var data Data
-			err := sc.Scan(&data)
-			if err != nil {
-				b.Fatal(err)
-			}
-			SumF64 += data.Event.F64
-		}
-	}
-	if err := sc.Err(); err != nil && err != io.EOF {
-		b.Fatal(err)
-	}
-}
-
-func TestChainSeekEntryTreeScannerPtr(t *testing.T) {
-	files := []string{
-		"../testdata/chain.flat.1.root",
-		"../testdata/chain.flat.2.root",
-	}
-
-	trees := make([]rtree.Tree, len(files))
-	for i, fname := range files {
-		f, err := riofs.Open(fname)
-		if err != nil {
-			t.Fatalf("could not open ROOT file %q: %v", fname, err)
-		}
-		defer f.Close()
-
-		obj, err := f.Get("tree")
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		trees[i] = obj.(rtree.Tree)
-	}
-	chain := rtree.Chain(trees...)
-
-	type Data struct {
-		F64    float64     `groot:"F64"`
-		ArrF64 [10]float64 `groot:"ArrF64"`
-		N      int32       `groot:"N"`
-		SliF64 []float64   `groot:"SliF64"`
-	}
-
-	sc, err := rtree.NewTreeScanner(chain, &Data{})
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer sc.Close()
-
-	var entries = []int64{0, 1, 4, 2, 5, 6, 7, 2, 7, 3, 0, 9}
-	for _, entry := range entries {
-		err := sc.SeekEntry(entry)
-		if err != nil {
-			t.Fatalf("Could not seek to entry %d: %v", entry, err)
-		}
-
-		if !sc.Next() {
-			t.Fatalf("Could not read entry %d", entry)
-		}
-
-		var data Data
-
-		err = sc.Scan(&data)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		i := sc.Entry()
-
-		if i != entry {
-			t.Fatalf("did not seek to entry %d. got=%d, want=%d", entry, i, entry)
-		}
-		if data.F64 != float64(i) {
-			t.Fatalf("entry [%d] : got= %#v want=%#v\n", i, data.F64, float64(i))
-		}
-		var arr [10]float64
-		for ii := range arr {
-			arr[ii] = float64(i)
-		}
-		if data.ArrF64 != arr {
-			t.Fatalf("entry [%d] : got= %#v want=%#v\n", i, data.ArrF64, arr)
-		}
-		sli := arr[:int(data.N)]
-		if !reflect.DeepEqual(sli, data.SliF64) {
-			t.Fatalf("entry [%d] : got= %#v want=%#v\n", i, data.SliF64, sli)
-		}
-	}
-}
-
-func TestChainSeekEntryTreeScannerVars(t *testing.T) {
-	files := []string{
-		"../testdata/chain.flat.1.root",
-		"../testdata/chain.flat.2.root",
-	}
-
-	trees := make([]rtree.Tree, len(files))
-	for i, fname := range files {
-		f, err := riofs.Open(fname)
-		if err != nil {
-			t.Fatalf("could not open ROOT file %q: %v", fname, err)
-		}
-		defer f.Close()
-
-		obj, err := f.Get("tree")
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		trees[i] = obj.(rtree.Tree)
-	}
-	chain := rtree.Chain(trees...)
-
-	type Data struct {
-		F64    float64     `groot:"F64"`
-		ArrF64 [10]float64 `groot:"ArrF64"`
-		N      int32       `groot:"N"`
-		SliF64 []float64   `groot:"SliF64"`
-	}
-
-	var data Data
-	rvars := []rtree.ReadVar{
-		{Name: "F64", Value: &data.F64},
-		{Name: "ArrF64", Value: &data.ArrF64},
-		{Name: "N", Value: &data.N},
-		{Name: "SliF64", Value: &data.SliF64},
-	}
-
-	sc, err := rtree.NewTreeScannerVars(chain, rvars...)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer sc.Close()
-
-	var entries = []int64{0, 1, 4, 2, 5, 6, 7, 2, 7, 3, 0, 9}
-	for _, entry := range entries {
-		err := sc.SeekEntry(entry)
-		if err != nil {
-			t.Fatalf("Could not seek to entry %d: %v", entry, err)
-		}
-
-		if !sc.Next() {
-			t.Fatalf("Could not read entry %d", entry)
-		}
-
-		var data Data
-
-		err = sc.Scan(&data.F64, &data.ArrF64, &data.N, &data.SliF64)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		i := sc.Entry()
-
-		if i != entry {
-			t.Fatalf("did not seek to entry %d. got=%d, want=%d", entry, i, entry)
-		}
-		if data.F64 != float64(i) {
-			t.Fatalf("entry [%d] : got= %#v want=%#v\n", i, data.F64, float64(i))
-		}
-		var arr [10]float64
-		for ii := range arr {
-			arr[ii] = float64(i)
-		}
-		if data.ArrF64 != arr {
-			t.Fatalf("entry [%d] : got= %#v want=%#v\n", i, data.ArrF64, arr)
-		}
-		sli := arr[:int(data.N)]
-		if !reflect.DeepEqual(sli, data.SliF64) {
-			t.Fatalf("entry [%d] : got= %#v want=%#v\n", i, data.SliF64, sli)
-		}
-	}
-}
-
-func TestChainSeekEntryScannerPtr(t *testing.T) {
-	files := []string{
-		"../testdata/chain.flat.1.root",
-		"../testdata/chain.flat.2.root",
-	}
-
-	trees := make([]rtree.Tree, len(files))
-	for i, fname := range files {
-		f, err := riofs.Open(fname)
-		if err != nil {
-			t.Fatalf("could not open ROOT file %q: %v", fname, err)
-		}
-		defer f.Close()
-
-		obj, err := f.Get("tree")
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		trees[i] = obj.(rtree.Tree)
-	}
-	chain := rtree.Chain(trees...)
-
-	type Data struct {
-		F64    float64     `groot:"F64"`
-		ArrF64 [10]float64 `groot:"ArrF64"`
-		N      int32       `groot:"N"`
-		SliF64 []float64   `groot:"SliF64"`
-	}
-
-	var data Data
-	sc, err := rtree.NewScanner(chain, &data)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer sc.Close()
-
-	var entries = []int64{0, 1, 4, 2, 5, 6, 7, 2, 7, 3, 0, 9}
-	for _, entry := range entries {
-		err := sc.SeekEntry(entry)
-		if err != nil {
-			t.Fatalf("Could not seek to entry %d: %v", entry, err)
-		}
-
-		if !sc.Next() {
-			t.Fatalf("Could not read entry %d", entry)
-		}
-
-		err = sc.Scan()
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		i := sc.Entry()
-
-		if i != entry {
-			t.Fatalf("did not seek to entry %d. got=%d, want=%d", entry, i, entry)
-		}
-		if data.F64 != float64(i) {
-			t.Fatalf("entry [%d] : got= %#v want=%#v\n", i, data.F64, float64(i))
-		}
-		var arr [10]float64
-		for ii := range arr {
-			arr[ii] = float64(i)
-		}
-		if data.ArrF64 != arr {
-			t.Fatalf("entry [%d] : got= %#v want=%#v\n", i, data.ArrF64, arr)
-		}
-		sli := arr[:int(data.N)]
-		if !reflect.DeepEqual(sli, data.SliF64) {
-			t.Fatalf("entry [%d] : got= %#v want=%#v\n", i, data.SliF64, sli)
-		}
-	}
-}
-
-func TestChainSeekEntryScannerVars(t *testing.T) {
-	files := []string{
-		"../testdata/chain.flat.1.root",
-		"../testdata/chain.flat.2.root",
-	}
-
-	trees := make([]rtree.Tree, len(files))
-	for i, fname := range files {
-		f, err := riofs.Open(fname)
-		if err != nil {
-			t.Fatalf("could not open ROOT file %q: %v", fname, err)
-		}
-		defer f.Close()
-
-		obj, err := f.Get("tree")
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		trees[i] = obj.(rtree.Tree)
-	}
-	chain := rtree.Chain(trees...)
-
-	type Data struct {
-		F64    float64     `groot:"F64"`
-		ArrF64 [10]float64 `groot:"ArrF64"`
-		N      int32       `groot:"N"`
-		SliF64 []float64   `groot:"SliF64"`
-	}
-
-	var data Data
-	rvars := []rtree.ReadVar{
-		{Name: "F64", Value: &data.F64},
-		{Name: "ArrF64", Value: &data.ArrF64},
-		{Name: "N", Value: &data.N},
-		{Name: "SliF64", Value: &data.SliF64},
-	}
-
-	sc, err := rtree.NewScannerVars(chain, rvars...)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer sc.Close()
-
-	var entries = []int64{0, 1, 4, 2, 5, 6, 7, 2, 7, 3, 0, 9}
-	for _, entry := range entries {
-		err := sc.SeekEntry(entry)
-		if err != nil {
-			t.Fatalf("Could not seek to entry %d: %v", entry, err)
-		}
-
-		if !sc.Next() {
-			t.Fatalf("Could not read entry %d", entry)
-		}
-
-		err = sc.Scan()
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		i := sc.Entry()
-
-		if i != entry {
-			t.Fatalf("did not seek to entry %d. got=%d, want=%d", entry, i, entry)
-		}
-		if data.F64 != float64(i) {
-			t.Fatalf("entry [%d] : got= %#v want=%#v\n", i, data.F64, float64(i))
-		}
-		var arr [10]float64
-		for ii := range arr {
-			arr[ii] = float64(i)
-		}
-		if data.ArrF64 != arr {
-			t.Fatalf("entry [%d] : got= %#v want=%#v\n", i, data.ArrF64, arr)
-		}
-		sli := arr[:int(data.N)]
-		if !reflect.DeepEqual(sli, data.SliF64) {
-			t.Fatalf("entry [%d] : got= %#v want=%#v\n", i, data.SliF64, sli)
-		}
+		t.Fatalf("entries differ: got=%d want=%d", total.got, total.want)
 	}
 }
